v0.1

+++ circle-detector.py
@@ -8,8 +8,6 @@ import numpy as np
 import cv2
 from random import shuffle
 from daemonize import Daemonize
-from http.server import HTTPServer, BaseHTTPRequestHandler
-from socketserver import ThreadingMixIn
 
 rez=picamera.PiResolution(round(2028/2.5),round(1520/2.5)).pad()
 cx=int(str(rez).split("x")[0])//2
@@ -26,33 +24,51 @@ def get_circles(img):
         if debug: print("Got at least one circle")
         shuffle(circles)
         circles = np.round(circles[0,:]).astype("int")
+        offset_x=offset_y=0
         for (x,y,r) in circles:
             cv2.circle(bw,(x,y),r,(0,0,255),4)
             offset_x=cx-x
             offset_y=cy-y
             if debug: cv2.circle(bw,(x+offset_x,y+offset_y),r,(255,255,0),2)
-            else: return([offset_x,offset_y])
         if debug: cv2.circle(bw,(cx,cy),5,(0,255,255),2)
-    else:
-        if debug:
-            print("Nothing found")
-            cv2.imwrite("output.jpg",np.hstack([image,bw]))
+        return([offset_x,offset_y])
+    elif debug:
+        print("Nothing found")
+        cv2.imwrite("output.jpg",np.hstack([image,bw]))
+        return([1000,1000])
+
+def yamlxy(x,y):
+    return("---\nx: "+str(x)+"\ny: "+str(y)+"\n")
+
+def init():
+    writef("status","ON\n")
+    writef("output",yamlxy(0,0))
+
+def writef(f,c):
+    with open("/dev/shm/detector_"+f+".run",'w') as file: file.write(c)
+
+def getf(f):
+    with open("/dev/shm/detector_"+f+".run",'r') as file: return(file.read().replace('\n',''))
 
 def main():
+    init()
     with picamera.PiCamera(resolution=rez,framerate=15,sensor_mode=2) as camera:
     	time.sleep(0.2)
     	with picamera.array.PiRGBArray(camera) as stream:
         	while True:
-            		if debug: start = time.time()
-            		camera.capture(stream,format='bgr',use_video_port=True)
-            		off=get_circles(stream.array)
-            		print(off)
-            		stream.truncate()
-            		stream.seek(0)
-            		if debug: print(time.time()-start)
+                    if getf("status")=="ON":
+                        if debug: start = time.time()
+                        camera.capture(stream,format='bgr',use_video_port=True)
+                        off=get_circles(stream.array)
+                        writef("output",yamlxy(off[0],off[1]))
+                        stream.truncate()
+                        stream.seek(0)
+                        if debug: print(time.time()-start)
+                    else:
+                        time.sleep(1)
 
 if len(sys.argv)==2 and sys.argv[1]=="-d":
-    daemon=Daemonize(app="vial-detector",pid=pid,action=main)
+    daemon=Daemonize(app="detector",pid=pid,action=main)
     daemon.start()
 else:
     main()
